requirements.txt
anyio==2.2.0
anykeystore==0.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
autoattack==0.1
Babel==2.9.0
backcall==0.2.0
backports.cached-property==1.0.2
backports.functools-lru-cache==1.6.1
bleach==3.3.0
bottle==0.12.18
brotlipy==0.7.0
CacheControl==0.12.11
cached-property==1.5.2
cachy==0.3.0
certifi==2020.12.5
cffi==1.13.2
chardet==4.0.0
cleo==1.0.0a5
click==7.1.2
crashtest==0.3.1
cryptacular==1.6.2
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.6
docopt==0.6.2
dulwich==0.20.50
easydict==1.9
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.8.0
Flask==1.1.2
grad-cam==1.4.6
greenlet==2.0.1
grip==4.5.2
h5py==3.1.0
html5lib==1.1
hupper==1.10.3
idna==2.10
imageio==2.9.0
-e git+https://github.com/hendrycks/robustness.git@305054464935cdf4c5f182f1387a1cc506854d49#egg=imagenet_c&subdirectory=ImageNet-C/imagenet_c
importlib-metadata==4.13.0
importlib-resources==5.10.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jaraco.classes==3.2.3
jedi==0.18.0
jeepney==0.8.0
Jinja2==2.11.3
joblib==0.14.1
json5==0.9.5
jsonschema==4.17.1
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyterlab==3.0.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
keyring==23.11.0
kiwisolver==1.1.0
lmdb==1.3.0
lockfile==0.12.2
Markdown==3.2.2
Markdown-Editor==1.0.4
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==9.0.0
msgpack==1.0.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.4
notebook==6.2.0
numpy==1.17.4
oauthlib==3.2.2
olefile==0.46
opencv-python==3.4.11.45
openpyxl==3.0.7
packaging==20.9
pandas==0.25.3
pandocfilters==1.4.2
parso==0.8.1
PasteDeploy==3.0.1
path-and-address==2.0.1
pbkdf2==1.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
pkginfo==1.8.3
pkgutil-resolve-name==1.3.10
plaster==1.1
plaster-pastedeploy==1.0.1
platformdirs==2.5.4
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.13.0
ptyprocess==0.7.0
pycparser==2.19
Pygments==2.8.0
pylev==1.4.0
pymdown-extensions==8.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==19.0.2
repoze.sendmail==4.4.1
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
scikit-image==0.19.2
scikit-learn==0.22
scipy==1.7.3
seaborn==0.9.0
SecretStorage==3.3.3
Send2Trash==1.5.0
shellingham==1.5.0
six==1.13.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.4.44
tensorboardX==2.1
terminado==0.9.2
testpath==0.4.4
thop==0.0.31.post2005241907
tifffile==2020.7.4
timm==0.4.12
tomlkit==0.11.6
torch==1.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.53.0
traitlets==5.0.5
transaction==3.0.1
translationstring==1.4
ttach==0.0.3
typing-extensions==3.7.4.3
urllib3==1.26.3
velruse==1.1.1
venusian==3.0.0
virtualenv==20.16.7
Wand==0.6.5
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xlrd==2.0.1
xmltodict==0.12.0
zipp==3.4.0
zope.deprecation==4.4.0
zope.interface==5.5.2
zope.sqlalchemy==1.6
